fix: conftest path isolation + loader torch_dtype param name
Two bugs:

1. tests/conftest.py (new): pytest runs in a subprocess that may cache
   system transformers before test imports. Add a conftest that inserts
   /kaggle/working/pkgs at sys.path[0] and evicts all transformers.*
   sys.modules entries at collection time, mirroring what the notebook
   kernel does. No-op outside Kaggle (guards on dir existence).

2. mini_vllm/model/loader.py: from_pretrained was called with dtype=
   (transformers 5.x API). transformers 4.46.3 uses torch_dtype=. Rename
   the kwarg so model loading works with our pinned version.

+++ mini_vllm/model/loader.py
@@ -47,7 +47,7 @@ def load_model(config: ModelConfig) -> tuple[nn.Module, Any]:
 
     model = AutoModelForCausalLM.from_pretrained(
         config.model_name_or_path,
-        dtype=torch_dtype,            # transformers 5.x renamed torch_dtype → dtype
+        torch_dtype=torch_dtype,
         trust_remote_code=config.trust_remote_code,
     )
     model = model.to(config.device)
